+++ src/main/java/conspire/powers/CubeRunePower.java
@@ -8,32 +8,33 @@ import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.PowerStrings;
 
-import basemod.interfaces.PostDrawSubscriber;
-
-public class CubeRunePower extends AbstractConspirePower implements PostDrawSubscriber {
+public class CubeRunePower extends AbstractConspirePower {
     public static final String POWER_ID = "conspire:CubeRune";
     private static final PowerStrings powerStrings = CardCrawlGame.languagePack.getPowerStrings(POWER_ID);
     public static final String NAME = powerStrings.NAME;
     public static final String[] DESCRIPTIONS = powerStrings.DESCRIPTIONS;
     public AbstractCreature source;
+    private boolean justApplied;
 
     public CubeRunePower(AbstractCreature owner, AbstractCreature source, int amount) {
         super(POWER_ID, NAME, owner);
         this.source = source;
         this.amount = amount;
         this.type = PowerType.DEBUFF;
+        this.justApplied = true;
         this.updateDescription();
     }
 
-    @Override
+    // Called from Conspire.receivePostDraw
     public void receivePostDraw(AbstractCard c) {
-        // TODO: interact with block from previous turns
-        this.flash();
-        AbstractDungeon.actionManager.addToBottom(new DamageAction(this.owner, new DamageInfo(this.source, this.amount, DamageInfo.DamageType.THORNS)));
+        if (!this.justApplied) {
+            this.flash();
+            AbstractDungeon.actionManager.addToBottom(new DamageAction(this.owner, new DamageInfo(this.source, this.amount, DamageInfo.DamageType.THORNS)));
+        }
     }
 
     public static void loseBlockReplacement() {
-        if (AbstractDungeon.player.hasPower(POWER_ID)) {
+        if (AbstractDungeon.player.hasPower(POWER_ID) && !((CubeRunePower)AbstractDungeon.player.getPower(POWER_ID)).justApplied) {
             int aboutToDraw = AbstractDungeon.player.gameHandSize;
             aboutToDraw = Math.min(aboutToDraw, 10 - AbstractDungeon.player.hand.size());
             AbstractDungeon.player.loseBlock(Math.max(0, AbstractDungeon.player.currentBlock - aboutToDraw));